changed user_install.ps1 / schtask
The user_installps1 / sch task shouldn't run for a user if they already have the source files in their home drive. right now, script just checks for folder, ideally there will be a more comprehensive check or use of registry so installation only runs once (for every new user to the system).

+++ AppName/user_install.ps1
@@ -6,15 +6,17 @@ param(
 ## Get Application name from target folder:
 $ApplicationName = $TargetFolder | Split-Path -Leaf
 
-Copy-Item -Path "$TargetFolder" -Destination "C:\Users\$TargetUser" -Recurse -Exclude "$ApplicationName.lnk", "user_install.ps1"
+if (-not (Test-Path "C:\Users\$TargetUser\$ApplicationName" -ErrorAction SilentlyContinue)) {
+    Copy-Item -Path "$TargetFolder" -Destination "C:\Users\$TargetUser" -Recurse -Exclude "$ApplicationName.lnk", "user_install.ps1"
 
-## Create start menu folder and shortcut:
-$shortcut_folder = "C:\Users\$TargetUser\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\$ApplicationNAme"
-if (-not (Test-Path $shortcut_folder -ErrorAction SilentlyContinue)) {
-    New-Item -Path $shortcut_folder -ItemType Directory | Out-null
-}
-
-@("C:\Users\$TargetUser\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\$ApplicationNAme\",
-    "C:\Users\$TargetUser\Desktop\") | % {
-    Copy-Item -Path "$TargetFolder\$ApplicationName.lnk" -Destination "$_"
+    ## Create start menu folder and shortcut:
+    $shortcut_folder = "C:\Users\$TargetUser\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\$ApplicationNAme"
+    if (-not (Test-Path $shortcut_folder -ErrorAction SilentlyContinue)) {
+        New-Item -Path $shortcut_folder -ItemType Directory | Out-null
+    }
+    
+    @("C:\Users\$TargetUser\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\$ApplicationNAme\",
+        "C:\Users\$TargetUser\Desktop\") | % {
+        Copy-Item -Path "$TargetFolder\$ApplicationName.lnk" -Destination "$_"
+    }
 }
